+++ api/tests/certification/session-management/unit/application/session-controller_test.js
@@ -73,4 +73,39 @@ describe('Unit | Controller | sessionController', function () {
       expect(result).to.equal(serializedJurySessionsForPaginatedList);
     });
   });
+
+  describe('#getJurySession', function () {
+    const sessionId = 123;
+
+    beforeEach(function () {
+      sinon.stub(usecases, 'getJurySession');
+
+      request = {
+        auth: { credentials: { userId } },
+        params: {
+          id: sessionId,
+        },
+      };
+    });
+
+    context('when session exists', function () {
+      it('should reply serialized session informations', async function () {
+        // given
+        const jurySessionSerializer = { serialize: sinon.stub() };
+        const foundJurySession = Symbol('foundSession');
+        const serializedJurySession = Symbol('serializedSession');
+        const hasSupervisorAccess = true;
+        usecases.getJurySession
+          .withArgs({ sessionId })
+          .resolves({ jurySession: foundJurySession, hasSupervisorAccess });
+        jurySessionSerializer.serialize.withArgs(foundJurySession, hasSupervisorAccess).resolves(serializedJurySession);
+
+        // when
+        const response = await sessionController.getJurySession(request, hFake, { jurySessionSerializer });
+
+        // then
+        expect(response).to.deep.equal(serializedJurySession);
+      });
+    });
+  });
 });
